requirements.txt
fastapi
uvicorn
deepface
opencv-python
pillow
requests